Transaction feedback on bridge
